Added delay function to card reset process...
This fixes the issue with card init failing if card was inserted at
boot. Now it should work in all circumstances.

Preset delay used for the time being but addtitional code can be added
to check status of SCFG_MC bit0 . This bit can be polled to determin if
a card is inserted. Thus a new function can be added to pause until a
card is inserted so users don't mistanly attempt to use NitroHax with no
card inserted.

fifo stuff sugged by ahezard. This is my implementation of it.

Also changed is re-enabling SD access (this was commented out by
mistsake in a previous commit from Shutterbug's fork that had not yet
been corrected here), updated version string to 0.93, and misc other
small improvements.