Cleaned up card reset code and locked bi31 before booting game.
Cleaned up card reset code and improved it further. Noticed that my
gateway blue cart wouln't init correctly if it was inserted while
NitroHax booted. (but would if I left it ejected first)

So some carts need a delay between power off and on.

Arm7 now waits inbetween power off and on. Arm9 controls the delay.
Refer to the code for how that works.

I also have the cheat engine on arm9 reset SCFG_EXT to 0x02000000 (arm9
can't touch bit31 else it wont work)

Then have main.arm7.c of bootloader reset SCFG_CLK to 0x80 (this is how
it is in unmodified NTR mode) and set bit31 of SCFG_EXT.

This succesfully restores SCFG lock down and should improve
compatibility with games that may break as a result of the new dev
launcher patch.