Major overhaul. Boot Splash Animation implemented
Max Mod audio system and boot splash animation from NTR Launcher merged
with this project.

This is now intended to launch in TWL mode. It will do a mode switch
into NTR mode prior to booting the game. Patched dev launcher is no
longer required to use this! :D

Also this build is now being compiled on this fork of libnds:

https://github.com/ahezard/libnds