runCheatEngine call moved to while loop...
Moved runCheatEngine function call in main.cpp to be inside while loop.

This is so that if an error occurs, a break occurs in the loop before
runCheatEngine is called. This allows the program to auto exit back to
home menu if the no cartridge inserted error screen is triggered.

The user can now press any button while on the error screen and it will
fade out and trigger the power off command. On 3DS, this command just
sends the user back to home menu as it's not possible to do a real power
down from legacy mode.